requirements_debug.txt
django-debug-toolbar==1.2.2
lxml
pytest==2.8.5
pytest-django==2.9.1
pyquery

#in case of error, install also
#py==1.4.31
